jarvis_server/requirements/requirements.txt
RPi.GPIO
rpi-rf
samsungctl
cec
python-magichue